exercises(dnd-character): stub, test: bikeshed types

+++ exercises/practice/dnd-character/dnd_character.zig
@@ -1,22 +1,22 @@
 pub const Character = struct {
-    strength: isize,
-    dexterity: isize,
-    constitution: isize,
-    intelligence: isize,
-    wisdom: isize,
-    charisma: isize,
-    hitpoints: isize,
+    strength: i8,
+    dexterity: i8,
+    constitution: i8,
+    intelligence: i8,
+    wisdom: i8,
+    charisma: i8,
+    hitpoints: i8,
 
     pub fn init() Character {
         @compileError("please implement the init method");
     }
 };
 
-pub fn modifier(n: isize) isize {
+pub fn modifier(n: i8) i8 {
     _ = n;
     @compileError("please implement the modifier function");
 }
 
-pub fn ability() isize {
+pub fn ability() i8 {
     @compileError("please implement the ability function");
 }
